refactor: Add babel support to unit test
