+++ package.json
@@ -50,20 +50,24 @@
   },
   "devDependencies": {
     "axios": "0.16.2",
-    "babel-cli": "6.24.1",
-    "babel-core": "6.25.0",
-    "babel-eslint": "7.2.3",
-    "babel-loader": "7.1.1",
+    "babel-cli": "^6.26.0",
+    "babel-core": "6.26.0",
+    "babel-eslint": "8.0.1",
+    "babel-loader": "7.1.2",
     "babel-plugin-flow-runtime": "0.11.1",
     "babel-plugin-transform-decorators-legacy": "1.3.4",
+    "babel-plugin-transform-es2015-classes": "^6.24.1",
     "babel-plugin-transform-runtime": "6.23.0",
     "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "1.5.2",
+    "babel-preset-env": "1.6.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-es2015-node4": "2.1.0",
     "babel-preset-flow": "6.23.0",
     "babel-preset-react": "6.24.1",
     "babel-preset-stage-2": "6.24.1",
     "babel-preset-stage-3": "6.24.1",
-    "babel-runtime": "6.23.0",
+    "babel-register": "^6.26.0",
+    "babel-runtime": "6.26.0",
     "codacy-coverage": "2.0.2",
     "codecov": "2.2.0",
     "coveralls": "2.13.1",
@@ -87,7 +91,7 @@
     "html-webpack-plugin": "2.29.0",
     "in-publish": "2.0.0",
     "localstorage-memory": "1.0.2",
-    "mocha": "3.4.2",
+    "mocha": "3.5.3",
     "mocha-lcov-reporter": "1.3.0",
     "node-sass": "4.5.3",
     "normalize.css": "7.0.0",
@@ -126,7 +130,7 @@
     "release": "standard-version -a -s",
     "prepublish": "in-publish && npm run build:webui || not-in-publish",
     "flow": "flow",
-    "test": "mocha ./test/functional ./test/unit --reporter=spec --full-trace",
+    "test": "BABEL_ENV=registry mocha --require babel-register ./test/functional ./test/unit --reporter=spec --full-trace",
     "pre:ci": "npm run build:webui",
     "test:ci": "npm run test:coverage",
     "test:only": "mocha ./test/functional ./test/unit",
@@ -135,9 +139,11 @@
     "coverage:publish": "nyc report --reporter=lcov | codecov",
     "lint": "eslint .",
     "lint:css": "stylelint 'src/**/*.scss' --syntax scss",
+    "dev:start": "BABEL_ENV=registry babel-node src/lib/cli",
+    "code:build": "BABEL_ENV=registry babel src/ --out-dir build/ --ignore src/webui/ --copy-files",
     "pre:webpack": "npm run lint && rimraf static/*",
     "dev:webui": "babel-node tools/dev.server.js",
-    "build:webui": "npm run pre:webpack && webpack --config tools/webpack.prod.config.babel.js",
+    "build:webui": "npm run pre:webpack && BABEL_ENV=ui webpack --config tools/webpack.prod.config.babel.js",
     "build:docker": "docker build -t verdaccio . --no-cache",
     "build:docker:rpi": "docker build -f Dockerfile.rpi -t verdaccio:rpi ."
   },
