+++ test/functional/index.js
@@ -18,6 +18,8 @@ describe('functional test verdaccio', function() {
       require('./lib/startup').start('./store/test-storage3', '/store/config-3.yaml'),
     ]).then(() => {
       done();
+    }).catch(function(error) {
+        console.error("error on start servers", error);
     });
 
   });
@@ -75,7 +77,7 @@ describe('functional test verdaccio', function() {
   require('./tags/addtag.spec')();
   require('./plugins/auth.spec')();
   require('./notifications/notify')();
-  // requires packages published to server1/server2
+  // // requires packages published to server1/server2
   require('./uplink.cache.spec')();
   require('./uplink.auth.spec')();
