+++ test/helper/register.js
@@ -1,0 +1,6 @@
+require("babel-polyfill");
+require('babel-core/register')({
+  ignore: /node_modules\/(?!ProjectB)/,
+  sourceMap: 'inline',
+}); 
+require('../../src/lib/cli');
