+++ test/helper/verdaccio-test
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+require("babel-register");
+require('../../src/lib/cli');
